fix: reuse auth session across steps and refactor test suite execution service
